+++ lib/anyway/dynamic_config.rb
@@ -12,9 +12,11 @@ module Anyway
       #   my_config = Anyway::Config.for(:my_app)
       #   # will load data from config/my_app.yml, secrets.my_app, ENV["MY_APP_*"]
       #
-      # TODO: add config_path option, env_prefix option
-      def for(name)
-        new(name: name, load: false).load_from_sources
+      def for(name, **options)
+        config = allocate
+        options[:env_prefix] ||= name.to_s.upcase
+        options[:config_path] ||= config.resolve_config_path(name, options[:env_prefix])
+        config.load_from_sources(name: name, **options)
       end
     end